+++ include/mpc/control/alternative.hpp
@@ -1,7 +1,7 @@
 /// @file alternative.hpp
 #pragma once
 #include <mpc/control/applicative.hpp>
-#include <mpc/prelude.hpp>
+#include <mpc/functional/perfect_forward.hpp>
 
 namespace mpc {
   // alternative
