+++ include/mpc/prelude/operations.hpp
@@ -1,7 +1,7 @@
 /// @file operations.hpp
 #pragma once
 #include <functional> // std::plus, etc.
-#include <mpc/functional.hpp> // perfect_forward
+#include <mpc/functional/perfect_forward.hpp>
 
 namespace mpc::inline cpo {
   // Arithmetic operations
